refactor(cli): extract session render bank to _session_render.py (#1331)
Move the six session render helpers (_use_notebook_table, _render_status,
_render_logout_outcome, _render_auth_check_result, _render_auth_inspect,
_render_auth_inspect_error) verbatim out of cli/session_cmd.py into a new
sibling cli/_session_render.py, importing them back so
register_session_commands keeps calling them through this module's own
namespace (preserving the patch surface).

session_cmd.py drops 1080 -> 752 LOC and comes off the module-size ratchet
allowlist; the new module is 357 LOC. No command moved, no public-API or
test-patch-surface change.
